+++ pkg/storage/chunk/chunk_store_test.go
@@ -890,9 +890,7 @@ func TestStore_DeleteChunk(t *testing.T) {
 	nonExistentChunk := dummyChunkForEncoding(model.Now(), metric3, encoding.Varbit, 200)
 
 	fooMetricNameMatcher, err := parser.ParseMetricSelector(`foo`)
-	if err != nil {
-		t.Fatal(err)
-	}
+	require.NoError(t, err)
 
 	for _, tc := range []struct {
 		name                           string